Retire deprecated CMS block types after data migration
Production page_blocks no longer reference the five domain-specific block types, so the renderer and editor can rely on the generic CMS blocks introduced by the migration. Historical migrations remain intact, while admin archive/collection management stays because those pages still use the domain data modules.

Constraint: Production DB prerequisite query returned deprecated_block_count=0 before implementation

Rejected: Remove backend archive/collection modules | admin archive and collection pages still depend on them

Confidence: high

Scope-risk: moderate

Directive: Do not re-add archive_nilo/archive_process/archive_artist/collection_clay/collection_shape without a fresh migration and DB compatibility plan

Tested: npm run build

Tested: npm run test:run

Tested: cd backend && npm run build && npm run test

Not-tested: backend e2e; no entity, schema, or migration changes
