+++ backend/src/modules/pages/__tests__/pages.service.spec.ts
@@ -275,19 +275,12 @@ describe('PagesService', () => {
       'archive_artist',
       'collection_clay',
       'collection_shape',
-    ])('아카이브/콜렉션 CMS 블록 타입 %s 생성을 허용한다', async (type) => {
+    ])('deprecated 아카이브/콜렉션 CMS 블록 타입 %s 생성을 거부한다', async (type) => {
       const dto = { type, content: { sectionTitle: '섹션 제목' } };
       mockPageRepository.findOne.mockResolvedValue({ id: 1 });
-      mockBlockRepository.create.mockReturnValue({ id: 1, ...dto, page_id: 1 });
-      mockBlockRepository.save.mockResolvedValue({ id: 1, ...dto, page_id: 1 });
 
-      const result = await service.createBlock(1, dto);
-
-      expect(result.type).toBe(type);
-      expect(mockBlockRepository.create).toHaveBeenCalledWith({
-        ...dto,
-        page_id: 1,
-      });
+      await expect(service.createBlock(1, dto)).rejects.toThrow(BadRequestException);
+      expect(mockBlockRepository.create).not.toHaveBeenCalled();
     });
