+++ backend/src/modules/pages/pages.service.ts
@@ -25,11 +25,6 @@ const SUPPORTED_BLOCK_TYPES = [
   'split_content',
   'brand_story',
   'journal_preview',
-  'archive_nilo',
-  'archive_process',
-  'archive_artist',
-  'collection_clay',
-  'collection_shape',
   'color_card_list',
   'timeline_list',
   'person_card_list',
