+++ src/app/[locale]/collection/page.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from 'next';
 import { getTranslations } from 'next-intl/server';
 import BlockRenderer from '@/components/shared/blocks/BlockRenderer';
-import { withCollectionBlockPrefetch } from '@/components/shared/blocks/serverPrefetch';
 import { fetchPage } from '@/lib/api-server';
 import type { Page } from '@/lib/api';
 
@@ -35,7 +34,7 @@ export default async function CollectionPage({ params }: CollectionPageProps) {
     );
   }
 
-  const blocks = await withCollectionBlockPrefetch((page as Page).blocks, locale);
+  const blocks = (page as Page).blocks;
 
   return (
     <div className="min-h-screen">
