+++ src/components/shared/admin/page-editor/__tests__/BlockPropertyPanel.test.tsx
@@ -4,43 +4,6 @@ import BlockPropertyPanel from '@/components/shared/admin/page-editor/BlockPrope
 import type { DraftBlock } from '@/components/shared/admin/page-editor/SortableBlockItem';
 
 describe('BlockPropertyPanel', () => {
-  it.each([
-    'archive_nilo',
-    'archive_process',
-    'archive_artist',
-    'collection_clay',
-    'collection_shape',
-  ] as const)('%s 블록의 섹션 헤딩 필드를 편집한다', (type) => {
-    const onUpdateContent = vi.fn();
-    const block: DraftBlock = {
-      id: 1,
-      type,
-      content: {
-        sectionLabel: '기존 라벨',
-        sectionTitle: '기존 제목',
-        sectionDesc: '기존 설명',
-      },
-      sort_order: 0,
-      is_visible: true,
-    };
-
-    render(<BlockPropertyPanel block={block} onUpdateContent={onUpdateContent} />);
-
-    fireEvent.change(screen.getByLabelText('섹션 제목'), {
-      target: { value: '새 제목' },
-    });
-
-    expect(onUpdateContent).toHaveBeenLastCalledWith(
-      1,
-      expect.objectContaining({
-        sectionLabel: '기존 라벨',
-        sectionTitle: '새 제목',
-        sectionDesc: '기존 설명',
-      }),
-    );
-  });
-
-
   it('색상 카드 리스트 항목을 추가하고 편집한다', () => {
     const onUpdateContent = vi.fn();
     const block: DraftBlock = {
