+++ src/components/shared/blocks/__tests__/BlockRenderer.test.tsx
@@ -77,79 +77,6 @@ vi.mock('@/lib/api', async () => {
     categoriesApi: {
       getTree: vi.fn().mockResolvedValue([]),
     },
-    archivesApi: {
-      getAll: vi.fn().mockResolvedValue({
-        niloTypes: [
-          {
-            id: 1,
-            name: 'Zhuni',
-            nameKo: '주니',
-            color: '#8b2f21',
-            region: '의흥',
-            description: '주니 설명',
-            characteristics: ['붉은 색'],
-            productUrl: '/products?clay=zhuni',
-            sortOrder: 0,
-            isActive: true,
-          },
-        ],
-        processSteps: [
-          {
-            id: 1,
-            step: 1,
-            title: '채토',
-            description: '산지에서 원토를 채굴',
-            detail: '채토 설명',
-          },
-        ],
-        artists: [
-          {
-            id: 1,
-            name: '진위명',
-            title: '국가급 공예미술사',
-            region: '의흥',
-            story: '장인 이야기',
-            specialty: '주니 서시호',
-            imageUrl: null,
-            productUrl: '/products?artist=jin',
-            sortOrder: 0,
-            isActive: true,
-          },
-        ],
-      }),
-    },
-    collectionsApi: {
-      getAll: vi.fn().mockResolvedValue({
-        clay: [
-          {
-            id: 1,
-            type: 'clay',
-            name: 'Zhuni',
-            nameKo: '주니',
-            color: '#8b2f21',
-            description: '주니 설명',
-            imageUrl: null,
-            productUrl: '/products?clay=zhuni',
-            sortOrder: 0,
-            isActive: true,
-          },
-        ],
-        shape: [
-          {
-            id: 2,
-            type: 'shape',
-            name: '서시',
-            nameKo: null,
-            color: null,
-            description: '서시 설명',
-            imageUrl: null,
-            productUrl: '/products?shape=xishi',
-            sortOrder: 1,
-            isActive: true,
-          },
-        ],
-      }),
-    },
   };
 });
 
@@ -225,23 +152,16 @@ describe('BlockRenderer', () => {
   });
 
   it.each([
-    ['archive_nilo', '보이차의 산지', '주니'],
-    ['archive_process', '제조 과정', '채토'],
-    ['archive_artist', '우리 장인들', '진위명'],
-    ['collection_clay', '다완 컬렉션', '주니'],
-    ['collection_shape', '형태 컬렉션', '서시'],
-  ] as const)('renders CMS section block type %s', async (type, title, itemText) => {
-    const blocks: PageBlock[] = [
-      makeBlock({
-        type,
-        content: { sectionTitle: title },
-      }),
-    ];
+    'archive_nilo',
+    'archive_process',
+    'archive_artist',
+    'collection_clay',
+    'collection_shape',
+  ] as const)('renders removed deprecated block type %s as unknown', (type) => {
+    render(<BlockRenderer blocks={[makeBlock({ type: type as PageBlock['type'], content: {} })]} />);
 
-    render(<BlockRenderer blocks={blocks} />);
-
-    expect(await screen.findByText(title)).toBeInTheDocument();
-    expect((await screen.findAllByText(itemText)).length).toBeGreaterThan(0);
+    expect(screen.getByText(/알 수 없는 블록 타입/)).toBeInTheDocument();
+    expect(screen.getByText(type)).toBeInTheDocument();
   });
