+++ src/lib/api/admin/archives.ts
@@ -1,5 +1,38 @@
 import { apiClient } from '../core';
-import type { NiloType, ProcessStep, Artist } from '../archives';
+
+export interface NiloType {
+  id: number;
+  name: string;
+  nameKo: string;
+  color: string;
+  region: string;
+  description: string;
+  characteristics: string[];
+  productUrl: string;
+  sortOrder: number;
+  isActive: boolean;
+}
+
+export interface ProcessStep {
+  id: number;
+  step: number;
+  title: string;
+  description: string;
+  detail: string;
+}
+
+export interface Artist {
+  id: number;
+  name: string;
+  title: string;
+  region: string;
+  story: string;
+  specialty: string;
+  imageUrl: string | null;
+  productUrl: string;
+  sortOrder: number;
+  isActive: boolean;
+}
 
 export interface CreateNiloTypeData {
   name: string;
