+++ src/lib/api/pages.ts
@@ -14,11 +14,6 @@ export interface PageBlock {
     | 'split_content'
     | 'brand_story'
     | 'journal_preview'
-    | 'archive_nilo'
-    | 'archive_process'
-    | 'archive_artist'
-    | 'collection_clay'
-    | 'collection_shape'
     | 'color_card_list'
     | 'timeline_list'
     | 'person_card_list'
